+++ bank-feeds/RELEASES.md
@@ -677,3 +677,13 @@ Based on:
 - [typescript v2.3.0] bank-feeds
 ### Releases
 - [NPM v2.3.0] https://www.npmjs.com/package/@codat/bank-feeds/v/2.3.0 - bank-feeds
+
+## 2023-11-13 01:29:39
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Bank-Feeds.yaml
+- Speakeasy CLI 1.119.0 (2.188.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.0.0] bank-feeds
+### Releases
+- [NPM v3.0.0] https://www.npmjs.com/package/@codat/bank-feeds/v/3.0.0 - bank-feeds